Parallelization implemented
