+++ quests/compression/compress.py
@@ -103,64 +103,3 @@ class DatasetCompressor:
         self._check_compression_method(method)
         return fps(self._descriptors, self._entropies, size, method=method)
     
-    def segment_compress(self,
-        num_sample: int,
-        num_chunks: int
-    ):
-        
-        ray.init(ignore_reinit_error=True)
-        
-        chunk_size = int(np.ceil(len(self._descriptors) / num_chunks))
-        
-        start_indexes = [i*chunk_size for i in range(num_chunks)]
-        sample_mini_size = [num_sample//num_chunks]*(num_chunks-1)
-        last_val = num_sample%num_chunks if num_sample%num_chunks != 0 else num_sample//num_chunks
-        sample_mini_size.append(last_val)
-        result_ids = [self.process_dataset.remote(self, self._descriptors[start:start+chunk_size],
-            self._entropies[start:start+chunk_size],
-            num_chunks=num_chunks,
-            num_sample=mini_sample) for start, mini_sample in zip(start_indexes, sample_mini_size)]
-        
-        results = ray.get(result_ids)
-        
-        ray.shutdown()
-
-        y = []
-        for i in range(len(results)):
-            y = np.concatenate([y, np.array(results[i]) + start_indexes[i]])
-        
-        return y
-    
-    @ray.remote
-    def process_dataset(
-        self, x: np.ndarray, initial_entropies: np.ndarray, num_chunks: int, num_sample: int):
-        
-        N = len(x)
-
-        if N <= num_sample:
-            return np.arange(N)
-
-        if N <= num_chunks * num_sample:
-            return fps(x, initial_entropies, size = num_sample, method='msc')
-
-        chunk_size = num_chunks * num_sample
-        num_subsets = int(np.ceil(N / chunk_size))
-        
-        y = []
-        for i in range(num_subsets):
-            start = i * chunk_size
-            chunk = x[start : start + chunk_size]
-            initial_entropies_chunk = initial_entropies[start : start + chunk_size]
-            y.append(
-                start
-                + np.array(
-                    fps(chunk, initial_entropies_chunk, size = num_sample, method = 'msc')
-                )
-            )
-        
-        y = np.concatenate(y)
-        result = []
-        for ind in y:
-            result.append(x[ind])
-        i = np.array(self.process_dataset.remote(self, result, initial_entropies[y], num_chunks, num_sample))
-        return y[ray.get(i)]
